0.7.6 All stats can be a decimal now(Just wont show in the screen) Free points per level can now be decimal(Make sure to put an f after the number in the config if you want it to be decimal)

+++ src/main/java/github/catchaos8/levelup/client/ClientStatData.java
@@ -1,17 +1,17 @@
 package github.catchaos8.levelup.client;
 
 public class ClientStatData {
-    private static int[] playerStats;
+    private static float[] playerStats;
 
-    public static void set(int[] stats) {
+    public static void set(float[] stats) {
         ClientStatData.playerStats = stats;
     }
 
-    public static int[] getPlayerStats() {
+    public static float[] getPlayerStats() {
         return playerStats;
     }
 
-    public int getStat(int type) {
+    public float getStat(int type) {
         return playerStats[type];
     }
 }
